libinst: factor out instrument interface
Move gpib.[ch] into libinst with intent
of refactoring later.
