+++ libinst/Makefile.am
@@ -1,0 +1,11 @@
+AM_CFLAGS = @GCCWARN@
+
+AM_CPPFLAGS = \
+	-I$(top_srcdir)
+	@GPIB_CPPFLAGS@
+
+noinst_LTLIBRARIES = libinst.la
+
+libinst_la_SOURCES = \
+	gpib.c \
+	gpib.h
