+++ libinst/gpib.c
