+++ libinst/gpib.h
